+++ src/arkanoid/board/Enemy.java
@@ -19,7 +19,7 @@ public class Enemy extends BaseObject {
     }
 
     public void render(Graphics g) {
-        g.drawImage(super.img, super.x, super.y, null);
+        g.drawImage(super.img, (int)super.x, (int)super.y, null);
     }
 
     public void tick() {
@@ -32,7 +32,6 @@ public class Enemy extends BaseObject {
             super.velX *= -1;
 
         collision();
-
     }
 
     private void collision() {
